move all components to `/components` dir
