+++ src/pages/Edit.jsx
@@ -1,10 +1,10 @@
 import { ValidateItemsStateProps } from "./../propValidations";
 import Header from "../components/Header";
-import ScrapItems from "../components/Edit/ScrapItems";
-import GrandTotal from "../components/HomePage/GrandTotal";
-import Footer from "./../components/Footer";
 import Aside from "./../components/Aside";
 import homeIcon from "./../assets/home.svg";
+import ScrapItemsEditEarning from "../components/ScrapItemsEditEarning";
+import GrandTotal from "../components/GrandTotal";
+import Footer from "./../components/Footer";
 
 Edit.propTypes = ValidateItemsStateProps;
